+++ packages/sdk-ui-ts/src/client/types/account.ts
@@ -1,5 +1,5 @@
 import { SubaccountTransfer } from '@injectivelabs/sdk-ts/dist/client/indexer/types/account'
-import { Token } from '@injectivelabs/token-metadata'
+import { TokenInfo } from '@injectivelabs/token-metadata'
 
 export interface UiSubaccountBalance {
   totalBalance: string
@@ -28,5 +28,5 @@ export interface UiSubaccountTransfer
 }
 
 export interface UiSubaccountTransferWithToken extends UiSubaccountTransfer {
-  token: Token
+  token: TokenInfo
 }